requirements.txt
# Polymarket API
py-clob-client>=0.17.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Configuration
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.27.0

# Templates
jinja2>=3.1.0

# Monitoring
prometheus-client>=0.21.0

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
